packages/prompttest/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompttest"
version = "1.0.0"
description = "Test framework for LLM prompt files"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [{ name = "Scott Converse" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "prompttools-core>=1.0,<2.0",
    "promptcost>=1.0,<2.0",
    "pydantic>=2.0",
    "typer[all]>=0.12",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
prompttest = "prompttest.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
]

[tool.hatch.build.targets.wheel]
packages = ["src/prompttest"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py39"
line-length = 100
